updated the official landing page and also included images and requested corrections
